test: add the CalendarPage view on the storybook, pass queryClientProvider to storybook
