--- .storybook/preview.ts
@@ -1,14 +1,0 @@
-import type { Preview } from '@storybook/nextjs'
-import '../app/globals.css'
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-       color: /(background|color)$/i,
-       date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
